add test case for anonymous default exported function circular reference and named default export function and fix
