+++ packages/core/integration-tests/test/html.js
@@ -2278,75 +2278,7 @@ describe('html', function () {
     let res = await run(b, {output: null}, {require: false});
     assert.deepEqual(await res.output, ['client', 'client', 'viewer']);
   });
-  it('should work for this case', async function () {
-    /*
-     * To reproduce error: run the following test
-     * Add an index.html file with the js file imported
-     * Start up localhost
-     * get "TypeError 0, _themedDefault.default) is not a function"
-     */
-    let b = await bundle(
-      path.join(__dirname, '/integration/circular-import-bug/index.js'),
-      {
-        outputFS: inputFS,
-        //mode: 'production', occurs in prod mode as well
-        defaultTargetOptions: {
-          shouldScopeHoist: false, // Does not occur with scope-hoisting on
-          shouldOptimize: false,
-          sourceMaps: false,
-        },
-      },
-    );
-    assertBundles(b, [
-      {name: 'index.html', type: 'html', assets: ['index.html']},
-      {
-        type: 'js',
-        assets: [
-          'somethingsomething',
-          'bundle-manifest.js',
-          'client.js',
-          'index.js',
-          'index.js',
-          'index.js',
-          'index.js',
-        ],
-      },
-      {type: 'js', assets: ['viewer.js']},
-    ]);
-  });
-  it.only('should work for this case (smaller)', async function () {
-    let b = await bundle(
-      path.join(
-        __dirname,
-        '/integration/circular-import-devpackager-bug/a.mjs',
-      ),
-      {
-        outputFS: inputFS,
-        //mode: 'production', occurs in prod mode as well
-        defaultTargetOptions: {
-          shouldScopeHoist: false, // Does not occur with scope-hoisting on
-          shouldOptimize: false,
-          sourceMaps: false,
-        },
-      },
-    );
-    assertBundles(b, [
-      {name: 'index.html', type: 'html', assets: ['index.html']},
-      {
-        type: 'js',
-        assets: [
-          'somethingsomething',
-          'bundle-manifest.js',
-          'client.js',
-          'index.js',
-          'index.js',
-          'index.js',
-          'index.js',
-        ],
-      },
-      {type: 'js', assets: ['viewer.js']},
-    ]);
-  });
+
   it('should not point to unrelated sibling bundles', async function () {
     await bundle(
       path.join(
