+++ packages/core/integration-tests/test/integration/circular-anonymous-default-export/a.mjs
