+++ packages/core/integration-tests/test/integration/circular-anonymous-default-export/c.mjs
@@ -2,4 +2,4 @@ import b1 from "./b.mjs";
 
 export const c = b1();
 
-console.log("c: b1 =", b1());
+result(c);
