--- packages/core/integration-tests/test/integration/circular-import-bug/esm/colors.js
@@ -1,8 +1,0 @@
-/* eslint-disable @atlaskit/design-system/ensure-design-token-usage */
-import themed from './utils/themed';
-
-export var background = themed();
-
-export var skeleton = function skeleton() {
-  return "something";
-};
