--- packages/core/integration-tests/test/integration/circular-import-bug/esm/constants.js
@@ -1,4 +1,0 @@
-import { skeleton as skeletonColor } from './colors'; //bug goes away without this
-export var SOME = 'THING';
-
-export { skeleton as skeletonColor };
