--- packages/core/integration-tests/test/integration/circular-import-bug/esm/utils/get-theme.js
@@ -1,4 +1,0 @@
-import { SOME } from '../constants'; //need this
-export default function getTheme(props) {
-   return 'a';
-}
