--- packages/core/integration-tests/test/integration/circular-import-bug/esm/utils/themed.js
@@ -1,10 +1,0 @@
-/* eslint-disable prefer-rest-params */
-
-import getTheme from './get-theme';
-
-export default function themed() {
-  return function (props) {
-    var a = getTheme(props);
-    return a;
-  };
-}
