--- packages/core/integration-tests/test/integration/circular-import-bug/index.js
@@ -1,2 +1,0 @@
-import { default as themed } from './esm/utils/themed';
-themed();
