--- packages/core/integration-tests/test/integration/circular-import-bug/package.json
@@ -1,3 +1,0 @@
-{
-    
-}
