+++ packages/core/integration-tests/test/integration/circular-named-default-export/a.mjs
@@ -1,0 +1,1 @@
+import "./b.mjs";
