+++ packages/core/integration-tests/test/integration/circular-named-default-export/c.mjs
@@ -1,0 +1,5 @@
+import b1 from "./b.mjs";
+
+export const c = b1();
+
+result(c);
